sonar: cleanup

+++ src/common/src/main/java/fr/riege/ebsl/common/pathfinding/goal/GoalBlock.java
@@ -8,9 +8,9 @@ public record GoalBlock(int x, int y, int z) implements Goal {
 
     @Override
     public double heuristic(int x, int y, int z) {
-        double dx = this.x - x;
-        double dy = this.y - y;
-        double dz = this.z - z;
+        double dx = (double) this.x - x;
+        double dy = (double) this.y - y;
+        double dz = (double) this.z - z;
         return Math.sqrt(dx * dx + dy * dy + dz * dz);
     }
