+++ src/common/src/main/java/fr/riege/ebsl/common/pathfinding/goal/GoalColumn.java
@@ -7,15 +7,15 @@ public record GoalColumn(int x, int z, double radius) implements Goal {
 
     @Override
     public boolean isInGoal(int x, int y, int z) {
-        double dx = this.x - x;
-        double dz = this.z - z;
+        double dx = (double) this.x - x;
+        double dz = (double) this.z - z;
         return Math.sqrt(dx * dx + dz * dz) <= radius;
     }
 
     @Override
     public double heuristic(int x, int y, int z) {
-        double dx = this.x - x;
-        double dz = this.z - z;
+        double dx = (double) this.x - x;
+        double dz = (double) this.z - z;
         return Math.max(0.0, Math.sqrt(dx * dx + dz * dz) - radius);
     }
